Create Secret Number
Create an algorithm that simulates the secret number game, where given a secret number between ( 0, 100 ) the player should try to guess what is the number. Every time the player tries to guess the number the algorithm should feedback him telling if the secret number is bigger or smaller than the one. If the player guesses the number the algorithm should congratulate the player.